requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
cycler==0.10.0
Cython==0.29.23
Django==3.2.8
gensim==4.1.2
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.2.1
sqlparse==0.4.2
threadpoolctl==3.0.0
urllib3==1.26.7
wordcloud==1.8.1
